requirements.txt
Pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.16.3
WeasyPrint==60.2
beautifulsoup4==4.12.3
PyPDF2==3.0.1